Refactor EventBus.
AbstractBus is now completely gone. RawBus is now
a concrete implementation (that actually has only
one additional method). 

Processor is now based on Runnables instead of 
Event<?> for EventBus. 

Response / reply patterns are now a bit more 
explicit, without re-setting keys. The intermediate
classes used for routing of events are removed.
